# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: APEX Racing Analyst API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runners:
    post:
      operationId: createRunner
      tags: [runners]
      summary: Add a runner to a race
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
  /runners/{id}:
    get:
      operationId: getRunner
      tags: [runners]
      summary: Get a runner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Runner detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
    patch:
      operationId: updateRunner
      tags: [runners]
      summary: Update a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
    delete:
      operationId: deleteRunner
      tags: [runners]
      summary: Delete / scratch a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: List APEX scores for a runner
      parameters:
        - name: runnerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: racecardId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Scores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApexScore"
    post:
      operationId: createScore
      tags: [scores]
      summary: Create or update APEX score for a runner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApexScoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApexScore"
  /scores/{id}:
    patch:
      operationId: updateScore
      tags: [scores]
      summary: Update APEX score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApexScoreUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApexScore"
...
  /upload/races:
    post:
      operationId: uploadRaces
      tags: [uploads]
      summary: Upload races CSV/spreadsheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
  /upload/results:
    post:
      operationId: uploadResults
      tags: [uploads]
      summary: Upload race results CSV/spreadsheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
...
components:
  schemas:
...
    Runner:
      type: object
      required: [id, racecardId, horseName, jockey, trainer, draw, weight]
      properties:
        id:
          type: integer
        racecardId:
          type: integer
        horseId:
          type: ["integer", "null"]
        horseName:
          type: string
        jockey:
          type: string
        trainer:
          type: string
        draw:
          type: ["integer", "null"]
        weight:
          type: string
        age:
          type: ["string", "null"]
        form:
          type: ["string", "null"]
        odds:
          type: ["string", "null"]
        isNonRunner:
          type: boolean
        scratched:
          type: boolean
        createdAt:
          type: string
    RunnerInput:
      type: object
      required: [racecardId, horseName, jockey, trainer, weight]
      properties:
        racecardId:
          type: integer
        horseId:
          type: ["integer", "null"]
        horseName:
          type: string
        jockey:
          type: string
        trainer:
          type: string
        draw:
          type: ["integer", "null"]
        weight:
          type: string
        age:
          type: string
        form:
          type: string
        odds:
          type: string
        isNonRunner:
          type: boolean
        scratched:
          type: boolean
    RunnerUpdate:
      type: object
      properties:
        horseName:
          type: string
        jockey:
          type: string
        trainer:
          type: string
        draw:
          type: ["integer", "null"]
        weight:
          type: string
        age:
          type: string
        form:
          type: string
        odds:
          type: string
        isNonRunner:
          type: boolean
        scratched:
          type: boolean
...
    ApexScore:
      type: object
      required: [id, runnerId, racecardId, abilityScore, paceFitScore, tacticalResilienceScore, groundTripScore, replayIntelligenceScore, hiddenValueScore, volatilityRisk, totalScore, confidenceClass]
      properties:
        id:
          type: integer
        runnerId:
          type: integer
        racecardId:
          type: integer
        abilityScore:
          type: number
          description: "0-100 Ability rating"
        paceFitScore:
          type: number
          description: "0-100 Pace fit"
        tacticalResilienceScore:
          type: number
          description: "0-100 Tactical resilience"
        groundTripScore:
          type: number
          description: "0-100 Ground/trip suitability"
        replayIntelligenceScore:
          type: number
          description: "0-100 Replay intelligence"
        hiddenValueScore:
          type: number
          description: "0-100 Hidden value"
        volatilityRisk:
          type: number
          description: "0-100 Volatility risk (lower is better)"
        totalScore:
          type: number
          description: "Computed composite APEX score"
        confidenceClass:
          type: string
          description: "best_of_day | top_rated_high_variance | hidden_value | replay_upgrade | no_bet"
        analystNotes:
          type: ["string", "null"]
        replayContext:
          type: ["string", "null"]
        contextualIntelligence:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ApexScoreInput:
      type: object
      required: [runnerId, racecardId, abilityScore, paceFitScore, tacticalResilienceScore, groundTripScore, replayIntelligenceScore, hiddenValueScore, volatilityRisk, confidenceClass]
      properties:
        runnerId:
          type: integer
        racecardId:
          type: integer
        abilityScore:
          type: number
        paceFitScore:
          type: number
        tacticalResilienceScore:
          type: number
        groundTripScore:
          type: number
        replayIntelligenceScore:
          type: number
        hiddenValueScore:
          type: number
        volatilityRisk:
          type: number
        confidenceClass:
          type: string
        analystNotes:
          type: string
        replayContext:
          type: string
        contextualIntelligence:
          type: string
    ApexScoreUpdate:
      type: object
      properties:
        abilityScore:
          type: number
        paceFitScore:
          type: number
        tacticalResilienceScore:
          type: number
        groundTripScore:
          type: number
        replayIntelligenceScore:
          type: number
        hiddenValueScore:
          type: number
        volatilityRisk:
          type: number
        confidenceClass:
          type: string
        analystNotes:
          type: string
        replayContext:
          type: string
        contextualIntelligence:
          type: string
...
    UploadInput:
      type: object
      required: [filename, data]
      properties:
        filename:
          type: string
        data:
          type: array
          items:
            type: object
    UploadResult:
      type: object
      required: [success, rowsProcessed, rowsInserted, errors]
      properties:
        success:
          type: boolean
        rowsProcessed:
          type: integer
        rowsInserted:
          type: integer
        errors:
          type: array
          items:
            type: string