# lib/api-zod/src/generated/types/runner.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface Runner {
  id: number;
  racecardId: number;
  /** @nullable */
  horseId?: number | null;
  horseName: string;
  jockey: string;
  trainer: string;
  /** @nullable */
  draw: number | null;
  weight: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  form?: string | null;
  /** @nullable */
  odds?: string | null;
  isNonRunner?: boolean;
  scratched?: boolean;
  createdAt?: string;
}