# lib/api-zod/src/generated/types/runnerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface RunnerUpdate {
  horseName?: string;
  jockey?: string;
  trainer?: string;
  /** @nullable */
  draw?: number | null;
  weight?: string;
  age?: string;
  form?: string;
  odds?: string;
  isNonRunner?: boolean;
  scratched?: boolean;
}