# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  success: boolean;
  rowsProcessed: number;
  rowsInserted: number;
  errors: string[];
}